#3057414 get metrics in background
