+++ src/AzureApi/TooManyRequestsException.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace AzureBillingExporter.AzureApi
+{
+    public class TooManyRequestsException : Exception
+    {
+        public TooManyRequestsException(string message) : base(message)
+        {
+        }
+    }
+}
