+++ src/Configuration/Metrics/CustomCollectorConfig.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+using YamlDotNet.Serialization;
+
+namespace AzureBillingExporter.Configuration.Metrics
+{
+    public class CustomCollectorConfig
+    {
+        [YamlMember(Alias = "metrics", ApplyNamingConventions = false)]
+        public List<MetricConfig> Metrics { get; set; }
+    }
+}
