+++ src/Cost/CostDataCache.cs
@@ -1,0 +1,77 @@
+using System;
+using System.Collections.Concurrent;
+using System.Collections.Generic;
+
+namespace AzureBillingExporter.Cost
+{
+    public class CostDataCache
+    {
+        const int CacheItemExpirationTimeInMinutes = 5;
+
+        private readonly ConcurrentDictionary<string, CacheItem> _costDataCache =
+            new ConcurrentDictionary<string, CacheItem>();
+
+        public List<CostResultRows> GetDailyCost()
+        {
+            return GetCostByKey("Daily");
+        }
+
+        public void SetDailyCost(List<CostResultRows> data)
+        {
+            SetCostByKey("Daily", data);
+        }
+
+        public List<CostResultRows> GetMonthlyCost()
+        {
+            return GetCostByKey("Monthly");
+        }
+
+        public void SetMonthlyCost(List<CostResultRows> data)
+        {
+            SetCostByKey("Monthly", data);
+        }
+
+        public List<CostResultRows> GetCostByKey(string key)
+        {
+            var cacheItem = GetCacheItem(key);
+            if (cacheItem == null)
+            {
+                return new List<CostResultRows>();
+            }
+
+            return cacheItem.Data;
+        }
+
+        public void SetCostByKey(string key, List<CostResultRows> data)
+        {
+            SetCacheItem(key, data);
+        }
+
+        private CacheItem GetCacheItem(string key)
+        {
+            var expireDate = DateTime.UtcNow.Add(-1*TimeSpan.FromMinutes(CacheItemExpirationTimeInMinutes));
+            var cacheItem = _costDataCache.ContainsKey(key) ? _costDataCache[key] : null;
+            if (cacheItem == null || cacheItem.Created <= expireDate)
+            {
+                return null;
+            }
+
+            return _costDataCache[key];
+        }
+
+        private void SetCacheItem(string key, List<CostResultRows> data)
+        {
+             _costDataCache[key] = new CacheItem
+             {
+                 Created = DateTime.UtcNow,
+                 Data = data
+             };
+        }
+
+        class CacheItem
+        {
+            public DateTime Created { get; set; }
+            public List<CostResultRows> Data { get; set; }
+        }
+    }
+}
