+++ src/Startup.cs
@@ -1,5 +1,8 @@
 using System;
 using AzureBillingExporter.AzureApi;
+using AzureBillingExporter.Configuration;
+using AzureBillingExporter.Cost;
+using AzureBillingExporter.PrometheusMetrics;
 using Dodo.HttpClientResiliencePolicies;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
@@ -19,7 +22,7 @@ namespace AzureBillingExporter
             Configuration = configuration;
         }
 
-        public IConfiguration Configuration { get; }
+        private IConfiguration Configuration { get; }
 
         // This method gets called by the runtime. Use this method to add services to the container.
         // For more information on how to configure your application, visit https://go.microsoft.com/fwlink/?LinkID=398940
@@ -29,19 +32,20 @@ namespace AzureBillingExporter
                 new Uri("https://management.azure.com"), "AzureBillingExporter");
             services.Configure<ApiSettings>(Configuration.GetSection("ApiSettings"));
             services.AddSingleton(resolver => resolver.GetRequiredService<IOptions<ApiSettings>>().Value);
-            services.AddSingleton<BillingQueryClient>();
-            services.AddSingleton<AzureCostManagementClient>();
-            services.AddSingleton<AzureBillingMetricsGrapper>();
-            services.AddSingleton<CostDataCache>();
+            services.AddSingleton<IAccessTokenFactory, AccessTokenFactory>();
+            services.AddSingleton<IAccessTokenProvider, AccessTokenProvider>();
 
-            services.AddSingleton<CustomCollectorConfiguration>(serviceProvider=>
+            services.AddSingleton(serviceProvider =>
             {
                 var customCollectorsFilePath = Configuration["CustomCollectorsFilePath"];
                 return new CustomCollectorConfiguration(customCollectorsFilePath);
             });
 
-            services.AddSingleton<IAccessTokenFactory, AccessTokenFactory>();
-            services.AddSingleton<IAccessTokenProvider, AccessTokenProvider>();
+            services.AddSingleton<BillingQueryClient>();
+            services.AddSingleton<AzureCostManagementClient>();
+            services.AddSingleton<MetricsUpdater>();
+            services.AddSingleton<CostDataCache>();
+            services.AddSingleton<CustomMetricsService>();
 
             services.AddHostedService(resolver =>
             {
@@ -60,10 +64,12 @@ namespace AzureBillingExporter
             {
                 var billingQueryClient = resolver.GetRequiredService<BillingQueryClient>();
                 var costDataCache = resolver.GetRequiredService<CostDataCache>();
+                var customCollectorConfiguration = resolver.GetRequiredService<CustomCollectorConfiguration>();
                 var logger = resolver.GetRequiredService<ILogger<BackgroundCostCollectorHostedService>>();
                 return new BackgroundCostCollectorHostedService(
                     billingQueryClient,
                     costDataCache,
+                    customCollectorConfiguration,
                     logger);
             });
         }
@@ -80,17 +86,11 @@ namespace AzureBillingExporter
 
             app.UseRouting();
 
-            var billingGrapper = app.ApplicationServices.GetService<AzureBillingMetricsGrapper>();
-            Metrics.DefaultRegistry.AddBeforeCollectCallback(async (cancel) =>
-                {
-                    await billingGrapper.DownloadFromApi(cancel);
-                });
+            var azureBillingMetricsGrabber = app.ApplicationServices.GetService<MetricsUpdater>();
+            Metrics.DefaultRegistry.AddBeforeCollectCallback(() => { azureBillingMetricsGrabber.Update(); });
 
             // ASP.NET Core 3 or newer
-            app.UseEndpoints(endpoints =>
-            {
-                endpoints.MapMetrics();
-            });
+            app.UseEndpoints(endpoints => { endpoints.MapMetrics(); });
         }
     }
 }
